+++ addon/string/index.js
@@ -6,5 +6,9 @@ export { default as decamelize } from './decamelize';
 export { default as indexOf } from './index-of';
 export { default as lastIndexOf } from './last-index-of';
 export { default as length } from './length';
+export { default as split } from './split';
 export { default as substr } from './substr';
 export { default as substring } from './substring';
+export { default as toLower } from './to-lower';
+export { default as toUpper } from './to-upper';
+export { default as underscore } from './underscore';
