+++ addon/string/split.js
@@ -1,4 +1,4 @@
-import { resolveKeys } from './utils';
+import { resolveKeys } from '../utils';
 
 export default function(...keys) {
   return resolveKeys(keys, (str, sep) => {
