+++ addon/string/to-lower.js
@@ -1,4 +1,4 @@
-import { normalizeString } from './utils';
+import { normalizeString } from '../utils';
 
 export default function(key) {
   return normalizeString(key, val => val.toLowerCase());
